feat: allow --version to be supplied without loading Qt libraries
Running the app with a `--version` option allows users to provide
more useful debugging reports, and lets us detect the baked-in
app version in e.g. CI environments.

But it's of less use if just requesting the version also loads all
Qt libraries -- they may be exactly what's buggy, and/or may be
missing in a CI environment.